agents/research_learner/requirements.txt
google-genai>=1.0.0
pyyaml>=6.0
httpx>=0.27.0
